+++ src/features/genre/ui/GenreContentSection.jsx
@@ -5,8 +5,7 @@ import { activeGenreState } from '../model/genreAtom';
 import useDevice from "../../../shared/hooks/useDevice";
 import ContentCard from "../../../shared/components/ContentCard";
 import { getGenreListAPI } from "../../genre/api/genreAPI";
-import { getRegionAPI } from "../../../features/location/api/getRegionAPI";
-import { getZoneAPI } from "../../../features/location/api/getZoneAPI";
+import { getRegionAPI, getZoneAPI } from "../../../features/location/api/locationAPI";
 import FilterImg from '../../../shared/assets/icons/genre/filter.svg';
 import PeopleFilter from '../../../shared/components/filter/PeopleFilter';
 import SortFilter from '../../../shared/components/filter/SortFilter';
