+++ src/features/level/ui/LevelContentSection.jsx
@@ -6,8 +6,7 @@ import useDevice from "../../../shared/hooks/useDevice.js";
 import ContentCard from "../../../shared/components/ContentCard.jsx";
 import { levels } from '../model/levelData.js';
 import { getLevelListAPI } from "../api/levelAPI.js";
-import { getRegionAPI } from "../../../features/location/api/getRegionAPI";
-import { getZoneAPI } from "../../../features/location/api/getZoneAPI";
+import { getRegionAPI, getZoneAPI } from "../../../features/location/api/locationAPI";
 import FilterImg from '../../../shared/assets/icons/genre/filter.svg';
 import PeopleFilter from '../../../shared/components/filter/PeopleFilter';
 import SortFilter from '../../../shared/components/filter/SortFilter';
