+++ src/features/location/ui/filter/LocationSortFilter.jsx
@@ -10,13 +10,13 @@ import {
   DropdownMenu, 
   DropdownHeader, 
   DropdownItem,
-  MenuWrapper } from "./LocationFilterStyles.js";
+  MenuWrapper } from "../../../../shared/styles/LocationFilterStyles.js";
 import Satisfaction from '../../../../shared/assets/icons/genre/satisfaction.svg';
 import DropDownImg from "../../../../shared/assets/icons/common/dropdown.svg";
 import Check from '../../../../shared/assets/icons/common/filterIcon/check.svg';
 import useDevice from "../../../../shared/hooks/useDevice.js";
 import useDropdown from "../../../../shared/hooks/useDropDown.js"; 
-import { sortOptions } from "./LocationOptionList.js";
+import { sortOptions } from "../../../../shared/components/filter/OptionList.js";
 import PropTypes from 'prop-types';
 
 const LocationSortFilter = forwardRef(({ onSelect, selected: externalSelected }, ref) => {
